Smoothed out the speed transitions + conversion int -> float

+++ Snowflake.pde
@@ -1,14 +1,14 @@
 class Snowflake{
 
-    int posX;
-    int posY;
-    int offset;
-    int speed;
-    int returnX;
-    int returnY;
+    float posX;
+    float posY;
+    float offset;
+    float speed;
+    float returnX;
+    float returnY;
     boolean firstDraw = true;
 
-    public Snowflake(int posX, int posY, int offset, int speed){
+    public Snowflake(float posX, float posY, float offset, float speed){
         this.posX = posX;
         this.posY = posY;
         this.offset = offset;
@@ -23,7 +23,7 @@ class Snowflake{
             returnV = 0;
         }
         else{
-            int rPosY;
+            float rPosY;
             if(posY >= height - 30){
                 if(posX >= mouseX - 150 && posX <= mouseX + 150){
                     returnV = 1;
@@ -47,7 +47,7 @@ class Snowflake{
         return returnV;
     }
 
-    public void createSFShape(int x, int y, int offset){
+    public void createSFShape(float x, float y, float offset){
         stroke(255);
         line(x, y - offset, x, y + offset);
         line(x - offset, y, x + offset, y);
@@ -57,15 +57,15 @@ class Snowflake{
         returnY = y;
     }
 
-    public int getX(){
+    public float getX(){
         return returnX;
     }
 
-    public int getY(){
+    public float getY(){
         return returnY;
     }
 
-    public void setSpeed(int speed){
+    public void setSpeed(float speed){
         this.speed = speed;
     }
 }
